pyproject.toml
[project]
name = "info-collector"
version = "0.1.0"
description = "Topic-driven YouTube information collection pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
info-collector = "info_collector.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/info_collector"]

[tool.hatch.build]
packages = ["src/info_collector"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
